refactor: AuthHelper로 인증 관련 코드 분리
