+++ backend/src/test/java/com/stampcrush/backend/api/manager/cafe/ManagerCafeCouponSettingCommandApiControllerTest.java
@@ -6,6 +6,7 @@ import com.stampcrush.backend.api.manager.cafe.request.CafeCouponSettingUpdateRe
 import com.stampcrush.backend.application.manager.cafe.ManagerCafeCouponSettingCommandService;
 import com.stampcrush.backend.entity.user.Owner;
 import com.stampcrush.backend.fixture.OwnerFixture;
+import com.stampcrush.backend.helper.AuthHelper.OwnerAuthorization;
 import com.stampcrush.backend.repository.user.OwnerRepository;
 import com.stampcrush.backend.repository.user.RegisterCustomerRepository;
 import org.junit.jupiter.api.Test;
@@ -14,10 +15,10 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.util.Base64;
 import java.util.List;
 import java.util.Optional;
 
+import static com.stampcrush.backend.helper.AuthHelper.createOwnerAuthorization;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
@@ -64,7 +65,7 @@ class ManagerCafeCouponSettingCommandApiControllerTest {
     void 카페_쿠폰_정책_변경_시_인증이_안되면_401_상태코드를_반환한다() throws Exception {
         OwnerAuthorization ownerAuthorization = createOwnerAuthorization(OwnerFixture.GITCHAN);
 
-        when(ownerRepository.findByLoginId(ownerAuthorization.owner.getLoginId()))
+        when(ownerRepository.findByLoginId(ownerAuthorization.getOwner().getLoginId()))
                 .thenReturn(Optional.empty());
 
         String requestBody = formatRequestBody(CAFE_COUPON_SETTING_UPDATE_REQUEST);
@@ -84,7 +85,7 @@ class ManagerCafeCouponSettingCommandApiControllerTest {
 
         OwnerAuthorization ownerAuthorization = createOwnerAuthorization(owner);
 
-        when(ownerRepository.findByLoginId(ownerAuthorization.owner.getLoginId()))
+        when(ownerRepository.findByLoginId(ownerAuthorization.getOwner().getLoginId()))
                 .thenReturn(Optional.of(owner));
 
         String requestBody = formatRequestBody(CAFE_COUPON_SETTING_UPDATE_REQUEST);
@@ -101,31 +102,5 @@ class ManagerCafeCouponSettingCommandApiControllerTest {
     private String formatRequestBody(CafeCouponSettingUpdateRequest request) throws JsonProcessingException {
         ObjectMapper objectMapper = new ObjectMapper();
         return objectMapper.writeValueAsString(request);
-    }
-
-    private OwnerAuthorization createOwnerAuthorization(Owner owner) {
-        String loginId = owner.getLoginId();
-        String encryptedPassword = owner.getEncryptedPassword();
-        String basicAuthHeader = "Basic " + Base64.getEncoder().encodeToString((loginId + ":" + encryptedPassword).getBytes());
-
-        return new OwnerAuthorization(owner, basicAuthHeader);
-    }
-
-    private final class OwnerAuthorization {
-        private final Owner owner;
-        private final String basicAuthHeader;
-
-        private OwnerAuthorization(Owner owner, String basicAuthHeader) {
-            this.owner = owner;
-            this.basicAuthHeader = basicAuthHeader;
-        }
-
-        public Owner getOwner() {
-            return owner;
-        }
-
-        public String getBasicAuthHeader() {
-            return basicAuthHeader;
-        }
     }
 }
